api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
tqdm==4.66.1
python-dotenv==1.0.1
gitpython==3.1.41 